refactor: renamed shared/ to common/
