+++ hals/src/common/command-result.ts
