+++ hals/src/common/number-range.ts
