+++ hals/src/common/page.ts
