+++ hals/src/common/param-map.ts
