+++ hals/src/common/result.ts
