+++ hals/src/common/time.ts
