+++ hals/src/common/validation.ts
